requirements-dev.txt
pytest>=8.0.0,<9
PyYAML>=6.0
